My booking list shown

+++ app/src/main/java/com/example/parkingnbeltran/domain/Space.kt
@@ -1,7 +1,7 @@
 package com.example.parkingnbeltran.domain
 
 data class Space(
-    val spaceId: Int,
+    val spaceId: Long,
     val type: SpaceType,
 )
