+++ app/src/main/java/com/example/parkingnbeltran/domain/Vehicle.kt
@@ -1,6 +1,6 @@
 package com.example.parkingnbeltran.domain
 
 data class Vehicle(
-    val vehicleId: Int,
+    val vehicleId: Long,
     val imgVehicle: String,
 )
